+++ CHANGES.md
@@ -2,9 +2,10 @@
 
 ## 0.7.0 (unreleased)
 
-- feat: context menu regular Open
+- feat: create new folder
 - feat: create text file within a folder
 - feat: "Open in Finder" for folders
+- feat: context menu regular Open
 - feat: hit Enter or use context menu to rename files or folders
 - fix: allow keyboard navigation on columns upon load
 - fix: position fixes for context menu